chore: move ExecutorchModule to types
