+++ src/types.ts
@@ -9,3 +9,20 @@ export interface Model {
   isModelReady: boolean;
   interrupt: () => void;
 }
+
+export type ETInput =
+  | Int8Array
+  | Int32Array
+  | BigInt64Array
+  | Float32Array
+  | Float64Array;
+
+
+export interface ExecutorchModule {
+  error: string | null;
+  isModelLoading: boolean;
+  isModelRunning: boolean;
+  forward: (input: ETInput, shape: number[]) => Promise<number[]>;
+  loadMethod: (methodName: string) => Promise<void>;
+  loadForward: () => Promise<void>;
+}
